:truck: chore(api) move updateLastQuestionState usecase to src
