+++ api/src/shared/domain/usecases/index.js
@@ -3,6 +3,8 @@ import { fileURLToPath } from 'node:url';
 
 import * as complementaryCertificationBadgeRepository from '../../../certification/complementary-certification/infrastructure/repositories/complementary-certification-badge-repository.js';
 import * as badgeRepository from '../../../evaluation/infrastructure/repositories/badge-repository.js';
+import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
+import * as challengeRepository from '../../infrastructure/repositories/challenge-repository.js';
 import { injectDependencies } from '../../infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../infrastructure/utils/import-named-exports-from-directory.js';
 const path = dirname(fileURLToPath(import.meta.url));
@@ -12,8 +14,10 @@ const usecasesWithoutInjectedDependencies = {
 };
 
 const dependencies = {
+  assessmentRepository,
   complementaryCertificationBadgeRepository,
   badgeRepository,
+  challengeRepository,
 };
 
 const sharedUsecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
