+++ api/src/shared/domain/usecases/update-last-question-state.js
@@ -1,5 +1,5 @@
-import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
+import { logger } from '../../infrastructure/utils/logger.js';
+import { Assessment } from '../models/Assessment.js';
 
 const updateLastQuestionState = async function ({
   assessmentId,
